Fix Inputs (#954)
* Make SteamworksClient construction fallible (#951)

When Steam is not running the creation of the steamworks::Client will
fail, and as-is that causes an unwrap in the constructor of
SteamworksClient. The way to work around it would be to create the
SteamworksClient manually, but this is not possible since the fields are
private. Therefore I'm suggesting to make the construction fallible so
you can choose to construct it yourself, and just not turn on the steam
server if the construction fails.

I've left the existing behaviour of the built-in fallback
SteamworksClient unwrap on initialization error, but maybe this should
be addressed too.

* fix

Fix logs

Fix host-server mode for avian_physics and spaceships example (#950)

* wip

* wip

* fix

* rust fmt

wip

---------

Co-authored-by: Ole Strøhm <Strohm99@gmail.com>