feat(rollback): add more conditions are rollback checks
Previously we were only checking for rollbacks when we received a new
confirmed state. Then we would compare the new state against the
prediction history to see if we should rollback.

We now support a RollbackPolicy that combines both state-checks (AR) and
input-checks.

State-Always -> rollback whenever we receive a new confirmed
state, without checking for mismatch. This would allow us to not store a
PredictionHistory.

State-Check -> what we do now

State-Disabled -> do not rollback check based on state. This could be
used for deterministic replication and lockstep.

Input-Always -> rollback whenever we receive a new remote input.
  TODO: check that the new input is more recent than the previous value
  we had

Input-Check -> rollback if we receive a new input that doesn't match
what was predicted.