feat(rollback): Add RollbackMode::Check for inputs
Previously we had RollbackMode::Always, which meant that whenever we
receive new input messages, we would rollback to the previous last
confirmed ticks for the inputs.

It is possible to reduce the number of rollbacks by checking if any of
the newly received inputs contain mismatches with our existing
InputBuffer. If they don't, no need to rollback!

This PR refactors the ActionStateSequence trait by adding a common
`update_buffer` method to update the input buffer from the action
sequence. The method now returns a `mismatch_tick` which is the earliest
tick where we have a mismatch. It is used to select the starting tick at
which we will rollback.