feat: introduce default ReplicationGroup
The ReplicationGroup feature was added to have a guarantee that updates
for multiple entities would be sent in the same message (and would
therefore be processed together).

This can be useful for:
- hierarchy: as we want all the entities in the hierarchy to be received
  together (this could be relaxed, as we want to spawn them together but
  afterwards we could split updates across messages)
- prediction: we want all the entities that are predicted to be on the
  same tick (this could also be relaxed, but it could cause more
  rollbacks)

Currently, if ReplicationGroup is not specified, each entity is
replicated in its own ReplicationGroup where the group_id is derived
from the entity. This is not efficient because:
- we have many small messages, each with their message headers, etc.
- we create many independent GroupChannels for replication
- we serialize the ReplicationGroupId every time, which takes 8 bytes

Instead for the default case where we don't care about the 'group'
property, we can use a default ReplicationGroupId (0) that doesn't
guarantee that updates for all entities in the group will be sent
together. Instead the updates will be packed into a single message up to
the MTU size. We will also serialize this ReplicationGroupId with a
single byte.

This PR also optimizes:
- we avoid storing intermediate versions of the updates/actions and
  instead serialize them directly into the final writer
- we optimize the serialization of Entity (and thus of
  ReplicationGroupId) to take fewer bytes

A TODO would be to add a benchmark that measure the amount of bytes used
for replication, to confirm that these changes reduce the amount of
bytes networked.